refactor: writer 수정
