+++ src/main/java/org/eternity/exam/MainApplication.java
@@ -1,6 +1,5 @@
 package org.eternity.exam;
 
-import org.eternity.exam.LectureReporter.StorageType;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -14,9 +13,8 @@ public class MainApplication {
         ConfigurableApplicationContext context = SpringApplication.run(MainApplication.class);
         JdbcClient jdbcClient = context.getBean(JdbcClient.class);
 
-        LectureReporter reporter = new LectureReporter(jdbcClient, new CsvFormatter());
+        LectureReporter reporter = new LectureReporter(new CsvFormatter(), new FileWriter());
         reporter.report(
-                StorageType.FILE,
                 new Lecture("객체지향 설계", LocalDate.of(2025, 9, 22), 3));
     }
 }
