+++ src/main/java/org/eternity/exam/StorageType.java
@@ -1,0 +1,3 @@
+package org.eternity.exam;
+
+public enum StorageType { DATABASE, FILE }
