+++ src/main/java/org/eternity/exam/Writer.java
@@ -1,5 +1,5 @@
 package org.eternity.exam;
 
 public interface Writer {
-
+    void write(String serialized) throws Exception;
 }
